Add requested review changes
