+++ lib/validation.js
@@ -65,5 +65,17 @@ function serialize (handle, data) {
   return handle[outputSchema](data)
 }
 
-module.exports = { build, validate, serialize }
+function isValidLogger (logger) {
+  var result = true
+  const methods = ['info', 'error', 'debug', 'fatal', 'warn', 'trace', 'child']
+  for (var i = 0; i < methods.length; i += 1) {
+    if (!logger[methods[i]] || typeof logger[methods[i]] !== 'function') {
+      result = false
+      break
+    }
+  }
+  return result
+}
+
+module.exports = { build, validate, serialize, isValidLogger }
 module.exports.symbols = { payloadSchema, querystringSchema, outputSchema, paramsSchema }
